+++ config/partitions_custom.csv
@@ -7,6 +7,6 @@
 
 nvs,        data,   nvs,     0x00b000,    0x002000,
 otadata,    data,   ota,     0x00d000,    0x002000,
-app0,       app,    ota_0,   0x010000,    0x190000,
-app1,       app,    ota_1,   0x1A0000,    0x190000,
-storage,    data,   spiffs,  0x330000,    0x0D0000 
+app0,       app,    ota_0,   0x010000,    0x1A0000,
+app1,       app,    ota_1,   0x1B0000,    0x1A0000,
+storage,    data,   spiffs,  0x350000,    0x0B0000
